Fix for ffmpeg.probe not returning values
Found a video which works but does not have any Video Detains (in right click -> properties -> details). This would cause the old version to crash as it trying to set frame_count and duration would cause a KeyError.
So when This occurs it now sets frame_count and duration = 0. And when updating status just prints the ffmpeg stderr output (Loses the "progress = XX.XX%") for the current render.